+++ src/qcommon/msg.cpp
@@ -644,7 +644,7 @@ delta functions with keys
 =============================================================================
 */
 
-int kbitmask[32] = {
+unsigned int kbitmask[32] = {
 	0x00000001, 0x00000003, 0x00000007, 0x0000000F,
 	0x0000001F, 0x0000003F, 0x0000007F, 0x000000FF,
 	0x000001FF, 0x000003FF, 0x000007FF, 0x00000FFF,
@@ -666,7 +666,7 @@ void MSG_WriteDeltaKey( msg_t *msg, int key, int oldV, int newV, int bits ) {
 
 int MSG_ReadDeltaKey( msg_t *msg, int key, int oldV, int bits ) {
 	if ( MSG_ReadBits( msg, 1 ) ) {
-		return MSG_ReadBits( msg, bits ) ^ ( key & kbitmask[bits] );
+		return MSG_ReadBits( msg, bits ) ^ static_cast<int>( static_cast<unsigned int>(key) & kbitmask[bits] );
 	}
 	return oldV;
 }
